price-intel-engine/requirements.txt
curl_cffi>=0.5.10
pydantic>=2.0.0
gspread>=5.10.0
oauth2client>=4.1.3
python-dotenv>=1.0.0
loguru>=0.7.0
curl_cffi>=0.5.10
pydantic>=2.0.0
pydantic-settings>=2.0.0
gspread>=5.10.0
oauth2client>=4.1.3
loguru>=0.7.0
tenacity>=8.2.3
pytest-asyncio>=0.23.0
